+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - improved level load times
 - improved logs module names readability
 - improved crash debug information on Windows
+- improved vertex movement when looking through water portals (#1493)
 
 ## [4.3](https://github.com/LostArtefacts/TR1X/compare/4.2...4.3) - 2024-08-15
 - added deadly water feature from TR2+ for custom levels (#1404)